Refactor: Break up monolithic main.ts
I've decomposed the main.ts file into smaller, more manageable modules:

- src/game.ts: Contains the main Phaser scene logic (Tetris class).
- src/tetronimo.ts: Contains the Tetronimo class definition.
- src/point.ts: Contains the Point class definition.
- src/constants.ts: Contains game-specific constants.

I updated src/main.ts to import the necessary components and initialize the Phaser game.
I also adjusted import paths across all affected files to reflect the new structure.
This refactoring improves code organization and maintainability.
