+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const INITIAL_INTERVAL = 500;
+export const INTERVAL_SPEEDUP = 0.8;
+export const INTERVAL_LINES = 10;
+
+export const TILT_THRESHOLD = 8;
